+++ packages/acme_chart/lib/src/core/interactive_layer/interactable_drawings/vertical_line/vertical_line_interactable_drawing.dart
@@ -187,6 +187,7 @@ class VerticalLineInteractableDrawing
         animationProgress: animationInfo.stateChangePercent,
         color: config.lineStyle.color,
         backgroundColor: chartTheme.backgroundColor,
+        utcOffset: chartConfig.chartTimeConfig.utcOffset,
         addNeonEffect: true,
       );
     }